Incorporate recent updates for Avalon & Innosilicon
