+++ pyasic/miners/_backends/__init__.py
@@ -17,3 +17,4 @@ from .bosminer import BOSMiner
 from .btminer import BTMiner
 from .cgminer import CGMiner
 from .hiveon import Hiveon
+from .cgminer_avalon import CGMiner_Avalon
